test(api): rewrite desktop auth bdd tests
